test:user request specを一つのファイルにまとめる
