+++ .rubocop_todo.yml
@@ -10,7 +10,7 @@ Metrics/AbcSize:
   Max: 30
 
 Metrics/BlockLength:
-  Max: 99
+  Max: 140
 
 Metrics/MethodLength:
   Max: 20
